+++ components/Comments/Comments.tsx
@@ -1,4 +1,4 @@
-import comments from "../data/comments.json";
+import comments from "../../data/comments.json";
 import Comment from "./Comment";
 // import { CommentInterface } from "../interfaces";
